Refactor install command
use main.rs and lib.rs setup like other commands
