+++ install/uvm-install/Cargo.toml
@@ -4,6 +4,8 @@ version = "2.0.1"
 authors = ["Manfred Endres <manfred.endres@tslarusso.de>"]
 
 [dependencies]
+flexi_logger = "0.9.2"
+log = "0.4.5"
 console = "0.6.1"
 serde = "1.0"
 serde_derive = "1.0"
